requirements.txt
attrs==19.3.0
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.0
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
gunicorn==19.9.0
importlib-metadata==0.23
jsonschema==3.1.1
more-itertools==7.2.0
psycopg2-binary==2.8.4
pyrsistent==0.15.5
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
zipp==0.6.0
